refactor: split expansion.c to convert_to_token_list.c
